requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
pycryptodome
Werkzeug